Use installed PHPUnit binary
Using Travis one is usefull when supporting different versions of PHP, as for example PHPUnit 5.2 is not supported on PHP 5.3 or 5.4. But in our case we don't really care and we are installing PHPUnit anyway, so we might as well use it directly instead of having to update Travis one.