[FEATURE] Pass controllerResult to SerializerContextBuilder
By passing the controllerResult to the SerializerContextBuilder we allow to control serializaton groups much more fine-grained.

This way we are able to control serialization groups not only based on corrently logged in user fx but rather also in combination with the actual controller result.

For example: When the current User is the author of some Product, add group „product_owner“