Document what count() is supposed to return
count() could return the number of pages, or the total number of items,
or the number of items on the current page.
Looking at the doctrine ORM implementation, it looks like it should be
the number of items on the current page, so let's document that, maybe
it is relied on somewhere?