Allow to deactivate fetch join collection
Having this option by default is a good choice since it offers better and more accurate perfomance in many situations. However there is some cases when having the fetch join collection counter productive : like many items retrieved for a simple query without join, as it forces to have 2 queries (first the disctinct, then an IN queries which is worst for performance in heavy table)

This commit allow users to set use fetch join collection to false (it still is true by default) in their resource / operation attributes for those special cases.