Avoid creating potentially crashing service
This is broken since sf 3.4.4, see https://github.com/symfony/symfony/pull/25823
One of api platforms listener depends on a file that now throws if it is
loaded and the expression language component is missing. The solution is
to avoid creating the service if the component is missing.
This fixes an error people will have even if they do not use the
access_control annotation attribute.