Fix fatal error when subresource type is not defined
Currently, when a property has a @ApiSubresource declaration, but no
proper type declaration, this results in a "Fatal error: Call to a
member function isCollection() on null", which is hard to debug.
Instead throw an exception referencing the offending property, so the
declaration can be fixed.

Additionally, add some unit tests for the
AnnotationSubresourceMetadataFactory.