Rework how resource class resolver use inheritance
This commits comes with many changes:

1. The resource class resolver now works well with inheritance and is
able to resolve resources using the ApiPlatform resource list (more
clever)
2. This change adds a new feature: inherited items from api resources
now are working as an api, a new behat test proves it
3. some parts of api platform were not compatible because it does not
uses the resource resolver but "get class" to get the resource, theses
parts are also fixed

Notice: the cache from the resource class resolver is removed here. It
was introduced in 2015 and was probably a great improvement, but since
2016 resources are not re-genarated on each call to the factory but we
call instead a cached factory. This cache was just adding complexity so
I removed it.