Make sure the order of nested order filters is preserved
Order filters are being written with underscores in GraphQL but are
sent along with dots to access nested entities.

The previous solution appended the dot version to the order array.
This leads to an issue where every order filter that accesses a nested
property will always be added to the query after local properties.