fix: Remove unnecessary duplicate select statement
This exact select statement is already added a few lines above. The only
exception is when fetchPartial is active, and in that case the current
implementation is wrong anyways, because it always adds the full select.

When this "temporary" solution for avoiding recursion was implemented,
the duplicated line above was not yet there:
https://github.com/api-platform/core/blob/5ba518014e2770e1ad5686b0124b2db45245fee5/src/Bridge/Doctrine/Orm/Extension/EagerLoadingExtension.php#L148

But later, the duplicated statement was added, and the select statement
inside the "Avoid recursion" case was made obsolete:
https://github.com/api-platform/core/commit/e34427a9f3e356a2d74350a035421dff6e499634