Paginated list without a list?!
When is a paginated list returned without the list or with a list of null values? But at the same time, pagination info is always there?!
Makes no sense to me, some information on the matter would be appreciated.

The same applies to cursor pagination - edges list can be a null? a node in an edge can be a null? How can that happen?!

My main issue is automatic type extraction for frontend integrations, where I constantly have to check if the list is there, if an item is actually an item and not a null, etc.

If there's a real possibility for these cases, can we somehow make this typing behaviour configurable instead?

Please share your thoughts, I am completely lost on this.