fix: support empty array as object
Fixes an issue with the use of `EMPTY_ARRAY_AS_OBJECT` and
`PRESERVE_EMPTY_OBJECTS` in the serialization context.
Since the CollectionNormalizer was taking over the serialization
of all iterables, the Symfony serializer was not
called when normalizing a "raw" collection.
Adding a better supports method fixes the issue.
The supports method is not cacheable anymore since it relies on the context.