fix(graphql): use default nested query operations
Instead of relying on a resource metadata factory to retrieve a "dynamic" nested query operation, this PR adds default nested query operations if the resource does not have one.
It simplifies a lot of things and catching the OperationNotFound exception is not needed anymore.
Since operations are "nested", they do not appear as top-level queries.
This PR also improves the XML/YAML compatibility.