do not add linked data for denormationContext
If I look in the generated schema for denormalizationContext i can see the required fields for '@id' and '@type'. Why do have this fields required for write operations? We don't need to pass this fields in a write request. If you used this schema on the client with typescript you need to write workarounds because you not sending this fields in the requests.

See https://github.com/api-platform/core/issues/5263