feat: add the ability to use previous_data object in state processor for update when Input DTO is set without the need to explicitly load the 'previous_data' again.
The issue has been raised in #4916 discussion.

Why we need this? Because 'previous_data' assigned in ReadListener is a clone of `data` and as a result it is not managed by the ObjectManager. If we would try to persist the modified object, instead of updating it the ObjectManager would try to perform a new insert and it will fail.