docs: explain how to disable the automatic integration with LexikJWTAuthenticationBundle
With #5609 the integration of API Platform with LexikJWTAuthenticationBundle is automatically enabled. However, if we are not using the default configuration of LexikJWTAuthenticationBundle the documented operation is wrong (wrong path, wrong response in my case) and creates a duplicate.
Since the behaviour of API Platform changed in a patch version maybe it's a good idea to explain directly in the changelog how to go back to the previous behaviour.