feat(pagination): provide a way to manually disable distinct on count
Using DISTINCT in a COUNT has a major impact on performance. This change
pushes the existing fix forward by delegating the responsibility to
disable the DISTINCT presence to the developers.

For instance, it is possible to safely remove the DISTINCT from a query
that joins a "to one" association.

The existing optimization still exists for most cases.