fix(metadata): compute isWritable during updates
Doctrine should not assume an `id` is not writable during updates.
PropertyInfo should not assume a property is not writable (for now only
during updates) but as this has a direct impact on serialization groups
maybe we should just skip readable/writable in the
PropertyInfoPropertyMetadataFactory.
SerializerPropertyMetadataFactory should run after AttributePropertyMetadataFactory or we can not override this behavior.

Mayabe that we could just simplify these and never set `writable` (or
`readable`) from Doctrine or PropertyInfo actually (and we won't need to
add the api_allow_update to the getFactoryOptions...).

fixes #7382